+++ mod/main.lua
@@ -6,6 +6,10 @@
 --[[
 
 TODO:
+- check to see if i can get safegridindex for devil room, boss rush
+- minimize w2 beam of light delay through testing
+- replace all current index and prev room index code
+- make timer smaller
 - add floor to starting screen
 - custom animation when you spawn close to mimic (make it red?)
 - remove character check from mod since we can restart X